+++ src/Sylius/Bundle/ShippingBundle/Calculator/FlatRateCalculator.php
@@ -11,7 +11,7 @@
 
 namespace Sylius\Bundle\ShippingBundle\Calculator;
 
-use Sylius\Bundle\ShippingBundle\Model\ShipmentInterface;
+use Sylius\Bundle\ShippingBundle\Model\ShippingSubjectInterface;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 
 /**
@@ -26,10 +26,8 @@ class FlatRateCalculator extends Calculator
      *
      * {@inheritdoc}
      */
-    public function calculate(ShipmentInterface $shipment)
+    public function calculate(ShippingSubjectInterface $subject, array $configuration)
     {
-        $configuration = $shipment->getMethod()->getConfiguration();
-
         return $configuration['amount'];
     }