+++ src/Sylius/Bundle/ShippingBundle/spec/Sylius/Bundle/ShippingBundle/Calculator/PerItemRateCalculator.php
@@ -11,16 +11,14 @@
 
 namespace spec\Sylius\Bundle\ShippingBundle\Calculator;
 
-use PHPSpec2\ObjectBehavior;
+use PhpSpec\ObjectBehavior;
 
 /**
- * Flat rate per item calculator spec.
- *
  * @author Paweł Jędrzejewski <pjedrzejewski@diweb.pl>
  */
-class PerItemRateCalculator extends ObjectBehavior
+class PerItemRateCalculatorSpec extends ObjectBehavior
 {
-    function it_should_be_initializable()
+    function it_is_initializable()
     {
         $this->shouldHaveType('Sylius\Bundle\ShippingBundle\Calculator\PerItemRateCalculator');
     }
@@ -30,7 +28,7 @@ class PerItemRateCalculator extends ObjectBehavior
         $this->shouldImplement('Sylius\Bundle\ShippingBundle\Calculator\CalculatorInterface');
     }
 
-    function it_should_be_configurable()
+    function it_is_configurable()
     {
         $this->shouldBeConfigurable();
     }
@@ -38,7 +36,7 @@ class PerItemRateCalculator extends ObjectBehavior
     /**
      * @param Symfony\Component\OptionsResolver\OptionsResolverInterface $resolver
      */
-    function it_should_have_required_amount_configuration_options($resolver)
+    function it_has_required_amount_configuration_options($resolver)
     {
         $resolver->setRequired(array('amount'))->shouldBeCalled()->willReturn($resolver);
         $resolver->setAllowedTypes(array('amount' => array('numeric')))->shouldBeCalled()->willReturn($resolver);
@@ -46,7 +44,7 @@ class PerItemRateCalculator extends ObjectBehavior
         $this->setConfiguration($resolver);
     }
 
-    function it_should_return_per_item_rate_configuration_form_type()
+    function it_returns_per_item_rate_configuration_form_type()
     {
         $this->getConfigurationFormType()->shouldReturn('sylius_shipping_calculator_per_item_rate_configuration');
     }